refactor(esm): finish the planner cell migration; restore the zero-warning build
Planner / converter:
- Carve the remaining cell-emission work out of PluginBuilder (-4,248 lines) into
  focused planner units: CellChildPlanBuilder (persistent / VWD / temporary
  bucketing plus the LAND + NAVM entries that lead the temporary bucket in
  vanilla file order), CellPlanFinalizer, PlacedRefLinkPlanner,
  PlanNavmConnectivity, PlanNavmEmission and OverrideSubrecordSanitizer —
  CellSectionPlanner had reached the 500-line architecture ceiling
- Add PluginEmissionCensus, which counts what a finished plugin actually
  contains by walking the assembled bytes. Per-emission counters cannot be
  trusted: records are written into buckets that later passes still discard
  wholesale, so counting at the write site over-reports
- Add typed runtime TREE reading and a SpeedTree record source; add
  MiscStaticObjectHandler coverage and RecordValidator checks
- Correct the FO3/FNV IMGS DNAM cinematic mask offset against retail bytes: the
  mask follows the body (132 when Skin Dimmer is present, else 128), and the
  four-float "Fade" block at cinematicBase + 32 does not exist — that region is
  the mask plus the record's uninitialized engine stack, so the old assertions
  pinned the implementation's own invention rather than the format

Rendering / viewer:
- Add a capture completion gate built on a demand-set census (CaptureSceneCensus):
  quiescence alone passes on the gap between streaming waves, and the demand set
  itself expands while streaming, so completion is treated as a fixpoint — all
  pending-work terms zero AND the demand-set terms unchanged across consecutive
  censuses
- Share one slope partition between the walk collision halves
  (WalkSurfaceSlopePolicy). The wall and ground thresholds deliberately overlap:
  a 45-60 degree face is both above step height and rest-able, and a gap would
  leave faces that are neither standable nor blocking
- Add walk void recovery, a placed-light cap, a material diffuse policy, and
  particle rate-controller coverage

GUI:
- Add PanelCollapseController (collapse a side panel to a strip, preserving the
  user's dragged splitter width), TextEntryFocusGuard and WinRtErrorInfo, plus
  a shared Core/Ui unit

Zero-warning build restored (46 unique warnings, 8 files):
- PluginBuilder: drop six dead private methods and their two now-orphaned
  helpers, four unread private fields, and two unused locals; demote
  _masterChildFormIds and _newWorldspacesForCellPipeline to locals in BuildAsync,
  their only method
- ReferenceRenderer12: keep the cull-veto chain intact behind a justified S3358
  pragma — the chain IS the gate order and flattening it loses the priority it
  encodes (matches the four existing S1244 pragmas in the file)
- Delete TopDownStreamingComplete, superseded in this batch by an inline
  StreamingQuiescence call that correctly passes null terrain stats when terrain
  depth is off, where the helper always dereferenced _terrain
- Make WarnOnImageSpaceLayoutVersionMismatch static; drop an unused CSV
  parameter; concretize WlstFormIdOffsets' return type; move a doc comment above
  its attributes; extract a nested ternary; justify an exact-zero float compare

Also updates two IMGS tests that still pinned the pre-correction 144/148 mask
offsets while the parser and their sibling had already moved to 132/128.

Full Release build: 0 errors, 0 warnings (both TFMs + companion apps).
Suite: 6,572 passed, 0 failed, 163 skipped with RUN_SHADER_COMPILE_TESTS=1.

+++ src/BethesdaMultitool/Core/Formats/Esm/Models/Records/Misc/TreeRecord.cs
@@ -38,8 +38,22 @@ public record TreeRecord
 }
 
 /// <summary>
-///     TREE CNAM payload (32 bytes, 8 LE floats). Matches the schema at
-///     <c>SubrecordCellAndMiscSchemas (CNAM, TREE, 32)</c>.
+///     TREE CNAM payload — 32 bytes: seven floats and one SIGNED INT32.
+///     <para>
+///     Field order and types are the engine's <c>OBJ_TREE</c> struct, read straight out of
+///     <c>Fallout_Release_MemDebug.pdb</c> (LF_FIELDLIST 0x0002dbf8, LF_STRUCTURE 0x0002dbf9,
+///     Size = 32), and they match xEdit's <c>wbStruct(CNAM)</c> one-for-one:
+///     </para>
+///     <list type="table">
+///         <item><description><c>+0  float  fCurveScalar</c>       → <see cref="LeafCurvature" /></description></item>
+///         <item><description><c>+4  float  fMinimumLeafAngle</c>  → <see cref="MinLeafAngle" /></description></item>
+///         <item><description><c>+8  float  fMaximumLeafAngle</c>  → <see cref="MaxLeafAngle" /></description></item>
+///         <item><description><c>+12 float  fBranchDimming</c>     → <see cref="BranchDimmingValue" /></description></item>
+///         <item><description><c>+16 float  fLeafDimming</c>       → <see cref="LeafDimmingValue" /></description></item>
+///         <item><description><c>+20 int32  iCanopyShadowRadius</c> → <see cref="ShadowRadius" /></description></item>
+///         <item><description><c>+24 float  fRockSpeed</c>         → <see cref="RockSpeed" /></description></item>
+///         <item><description><c>+28 float  fRustleSpeed</c>       → <see cref="RustleSpeed" /></description></item>
+///     </list>
 /// </summary>
 public record TreeData
 {
@@ -48,14 +62,23 @@ public record TreeData
     public float MaxLeafAngle { get; init; }
     public float BranchDimmingValue { get; init; }
     public float LeafDimmingValue { get; init; }
-    public float ShadowRadius { get; init; }
+
+    /// <summary>
+    ///     Canopy shadow radius in cells. INTEGER, not float — the PDB types this
+    ///     <c>iCanopyShadowRadius</c> as <c>T_INT4</c> and xEdit as <c>itS32</c>. Observed
+    ///     values are 128 / 512 / 64; decoding those bytes as a float yields ~7e-43 denormals,
+    ///     which is how the original float typing went unnoticed.
+    /// </summary>
+    public int ShadowRadius { get; init; }
+
     public float RockSpeed { get; init; }
     public float RustleSpeed { get; init; }
 }
 
 /// <summary>
-///     TREE BNAM payload (8 bytes, 2 LE floats). Matches the schema at
-///     <c>SubrecordCellAndMiscSchemas (BNAM, TREE, 8)</c>.
+///     TREE BNAM payload — 8 bytes, two floats. The engine's <c>NiPoint2 BillboardSize</c>
+///     (PDB LF_CLASS 0x000116bc, Size = 8, members <c>x</c>@0 and <c>y</c>@4), which xEdit
+///     names 'Billboard Dimensions' (Width, Height).
 /// </summary>
 public record TreeBillboardSize
 {
